move omp iterate functions to omp_kernel_launcher, some cleanup
