+++ include/hipSYCL/glue/omp/omp_kernel_launcher.hpp
@@ -30,7 +30,6 @@
 
 
 #include <cassert>
-#include <clang/Basic/OperatorKinds.h>
 #include <omp.h>
 
 #include "hipSYCL/common/debug.hpp"
@@ -57,6 +56,72 @@ namespace glue {
 namespace omp_dispatch {
 
 
+template <int Dim, class Function>
+void iterate_range_omp_for(sycl::range<Dim> r, Function f) {
+
+  if constexpr (Dim == 1) {
+    #pragma omp for
+    for (std::size_t i = 0; i < r.get(0); ++i) {
+      f(sycl::id<Dim>{i});
+    }
+  } else if constexpr (Dim == 2) {
+    #pragma omp for collapse(2)
+    for (std::size_t i = 0; i < r.get(0); ++i) {
+      for (std::size_t j = 0; j < r.get(1); ++j) {
+        f(sycl::id<Dim>{i, j});
+      }
+    }
+  } else if constexpr (Dim == 3) {
+    #pragma omp for collapse(3)
+    for (std::size_t i = 0; i < r.get(0); ++i) {
+      for (std::size_t j = 0; j < r.get(1); ++j) {
+        for (std::size_t k = 0; k < r.get(2); ++k) {
+          f(sycl::id<Dim>{i, j, k});
+        }
+      }
+    }
+  }
+}
+
+template <int Dim, class Function>
+void iterate_range_omp_for(sycl::id<Dim> offset, sycl::range<Dim> r,
+                           Function f) {
+
+  const std::size_t min_i = offset.get(0);
+  const std::size_t max_i = offset.get(0) + r.get(0);
+
+  if constexpr (Dim == 1) {
+  #pragma omp for
+    for (std::size_t i = min_i; i < max_i; ++i) {
+      f(sycl::id<Dim>{i});
+    }
+  } else if constexpr (Dim == 2) {
+    const std::size_t min_j = offset.get(1);
+    const std::size_t max_j = offset.get(1) + r.get(1);
+
+  #pragma omp for collapse(2)
+    for (std::size_t i = min_i; i < max_i; ++i) {
+      for (std::size_t j = min_j; j < max_j; ++j) {
+        f(sycl::id<Dim>{i, j});
+      }
+    }
+  } else if constexpr (Dim == 3) {
+    const std::size_t min_j = offset.get(1);
+    const std::size_t min_k = offset.get(2);
+    const std::size_t max_j = offset.get(1) + r.get(1);
+    const std::size_t max_k = offset.get(2) + r.get(2);
+
+  #pragma omp for collapse(3)
+    for (std::size_t i = min_i; i < max_i; ++i) {
+      for (std::size_t j = min_j; j < max_j; ++j) {
+        for (std::size_t k = min_k; k < max_k; ++k) {
+          f(sycl::id<Dim>{i, j, k});
+        }
+      }
+    }
+  }
+}
+
 template<class Function>
 inline 
 void single_task_kernel(Function f) noexcept
@@ -73,7 +138,7 @@ void parallel_for_kernel(Function f,
 
 #pragma omp parallel
   {
-    host::iterate_range_omp_for(execution_range, [&](sycl::id<Dim> idx) {
+    iterate_range_omp_for(execution_range, [&](sycl::id<Dim> idx) {
       auto this_item = 
         sycl::detail::make_item<Dim>(idx, execution_range);
 
@@ -93,7 +158,7 @@ void parallel_for_kernel(Function f,
 
 #pragma omp parallel
   {
-    host::iterate_range_omp_for(offset, execution_range, [&](sycl::id<Dim> idx) {
+    iterate_range_omp_for(offset, execution_range, [&](sycl::id<Dim> idx) {
       auto this_item = 
         sycl::detail::make_item<Dim>(idx, execution_range, offset);
 
@@ -110,7 +175,7 @@ inline void parallel_for_ndrange_kernel(
     size_t num_local_mem_bytes) noexcept
 {
 
-#ifdef HIPSYCL_NO_FIBERS
+#ifndef HIPSYCL_HAS_FIBERS
   rt::register_error(__hipsycl_here(),
                      rt::error_info{
                          "nd_range parallel for on CPU requires fibers, but 
@@ -161,7 +226,7 @@ void parallel_for_workgroup(Function f,
     sycl::detail::host_local_memory::request_from_threadprivate_pool(
         num_local_mem_bytes);
 
-    host::iterate_range_omp_for(num_groups, [&](sycl::id<Dim> group_id) {
+    iterate_range_omp_for(num_groups, [&](sycl::id<Dim> group_id) {
       sycl::group<Dim> this_group{group_id, local_size, num_groups};
       f(this_group);
     });
@@ -185,7 +250,7 @@ void parallel_region(Function f,
     sycl::detail::host_local_memory::request_from_threadprivate_pool(
         num_local_mem_bytes);
 
-    host::iterate_range_omp_for(num_groups, [&](sycl::id<dimensions> group_id) {
+    iterate_range_omp_for(num_groups, [&](sycl::id<dimensions> group_id) {
       sycl::group<dimensions> this_group{group_id, group_size, num_groups};
 
       auto phys_item = sycl::detail::make_sp_item(
@@ -216,13 +281,13 @@ public:
     
     this->_type = type;
 
-#ifdef HIPSYCL_NO_FIBERS
+#ifndef HIPSYCL_HAS_FIBERS
     if (type == rt::kernel_type::ndrange_parallel_for) {
       this->_invoker = []() {};
       
       throw sycl::feature_not_supported{
-        "The CPU backend does not support nd_range kernels because they cannot be implemented\n"
-        "efficiently in pure-library SYCL implementations such as hipSYCL on CPU. We recommend:\n"
+        "Support for nd_range kernels on CPU is disabled without fibers because they cannot be\n"
+        "implemented efficiently in pure-library SYCL implementations such as hipSYCL on CPU. We recommend:\n"
         " * to verify that you really need the features of nd_range parallel for.\n"
         "   If you do not need local memory, use basic parallel for instead.\n"
         " * users targeting SYCL 1.2.1 may use hierarchical parallel for, which\n"
@@ -230,6 +295,7 @@ public:
         "   and/or other SYCL implementations.\n"
         " * if you use hipSYCL exclusively, you are encouraged to use scoped parallelism:\n"
         "   https://github.com/illuhad/hipSYCL/blob/develop/doc/scoped-parallelism.md"
+        " * if you absolutely need nd_range parallel for, enable fiber support in hipSYCL."
       };
     }
 #endif
